+++ .github/workflows/ant.yml
@@ -5,9 +5,9 @@ name: Java CI
 
 on:
   push:
-    branches: [ "main" ]
+    branches: [ "*" ]
   pull_request:
-    branches: [ "main" ]
+    branches: [ "*" ]
 
 jobs:
   build:
